gizmo changes
