+++ packages/tools/guiEditor/src/diagram/gizmoWrapper.tsx
@@ -2,7 +2,9 @@ import type { Nullable } from "core/types";
 import type { Observer } from "core/Misc/observable";
 import * as React from "react";
 import type { GlobalState } from "../globalState";
-import { GuiGizmoComponent } from "./guiGizmo";
+import { GizmoGeneric } from "./gizmoGeneric";
+import { GizmoLine } from "./gizmoLine";
+import { Line } from "gui/2D/controls/line";
 
 export interface IGizmoWrapperProps {
     globalState: GlobalState;
@@ -22,9 +24,13 @@ export class GizmoWrapper extends React.Component<IGizmoWrapperProps> {
         const controls = this.props.globalState.selectedControls;
         return (
             <>
-                {controls.map((control) => (
-                    <GuiGizmoComponent globalState={this.props.globalState} control={control} key={control.uniqueId} />
-                ))}
+                {controls.map((control) =>
+                    control.getClassName() === "Line" ? (
+                        <GizmoLine globalState={this.props.globalState} control={control as Line} key={control.uniqueId} />
+                    ) : (
+                        <GizmoGeneric globalState={this.props.globalState} control={control} key={control.uniqueId} />
+                    )
+                )}
             </>
         );
     }
